+++ lib/features/about/about_constants.dart
@@ -1,0 +1,6 @@
+const buildMetadata = String.fromEnvironment(
+  'BUILD_METADATA',
+  defaultValue: '',
+);
+
+const projectHomeUrl = 'https://github.com/wisnij/unitary';